# swagger/v5/swagger.yaml
...
openapi: 3.0.1
info:
  title: API V5
  description: |-
    # Check financial eligibility for legal aid.

    ## Usage:
      - Create an assessment by POSTing a payload to `/assessments`
        and store the `assessment_id` returned.
      - Add assessment components, such as applicant, capitals and properties using the
        `assessment_id` from the first call
      - Retrieve the result using the GET `/assessments/{assessment_id}`
  contact:
    name: Github repository
    url: https://github.com/ministryofjustice/check-financial-eligibility
  version: v5
components:
  schemas:
...
    ProceedingTypeResult:
      type: object
      required:
      - ccms_code
      - client_involvement_type
      - upper_threshold
      - lower_threshold
      - result
      properties:
        ccms_code:
          type: string
          enum:
          - DA001
          - DA002
          - DA003
          - DA004
          - DA005
          - DA006
          - DA007
          - DA020
          - SE003
          - SE004
          - SE013
          - SE014
          - SE007
          - SE008
          - SE015
          - SE016
          - SE095
          - SE097
          - SE003A
          - SE004A
          - SE007A
          - SE008A
          - SE013A
          - SE014A
          - SE015A
          - SE016A
          - SE095A
          - SE097A
          - SE101A
          - IM030
          - IA031
          - SE003E
          - SE004E
          - SE007E
          - SE008E
          - SE013E
          - SE014E
          - SE015E
          - SE016E
          - SE096E
          - SE099E
          - SE100E
          - SE101E
          description: The code expected by CCMS
        client_involvement_type:
          type: string
          enum:
          - A
          - D
          - W
          - Z
          - I
          example: A
          description: The client_involvement_type expected by CCMS
        upper_threshold:
          type: number
        lower_threshold:
          type: number
        result:
          type: string
          enum:
          - eligible
          - ineligible
          - contribution_required
    PropertyResult:
      type: object
      additionalProperties: false
      properties:
        value:
          type: number
          minimum: 0.0
        outstanding_mortgage:
          type: number
          minimum: 0.0
        percentage_owned:
          type: integer
          minimum: 0
          maximum: 100
        main_home:
          type: boolean
        shared_with_housing_assoc:
          type: boolean
        transaction_allowance:
          type: number
          minimum: 0.0
        allowable_outstanding_mortgage:
          type: number
          minimum: 0.0
        net_value:
          type: number
        net_equity:
          type: number
        smod_allowance:
          type: number
          description: Amount of subject matter of dispute disregard applied to this
            property
          minimum: 0.0
          maximum: 100000.0
        main_home_equity_disregard:
          type: number
          description: Amount of main home equity disregard applied to this property
        assessed_equity:
          type: number
          minimum: 0.0
...
    Capitals:
      type: object
      additionalProperties: false
      properties:
        bank_accounts:
          "$ref": "#/components/schemas/BankAccounts"
        non_liquid_capital:
          type: array
          description: An array of objects describing applicant's non-liquid capital
            items (excluding property), e.g. valuable items, jewellery, trusts, other
            investments
          example:
          - value: 1.01
            description: asset name 1
            subject_matter_of_dispute: false
          - value: 100.01
            description: asset name 2
            subject_matter_of_dispute: true
          items:
            type: object
            description: Asset detail
            required:
            - value
            - description
            additionalProperties: false
            properties:
              value:
                "$ref": "#/components/schemas/positive_currency"
              description:
                description: Definition of a non-liquid capital item
                type: string
              subject_matter_of_dispute:
                description: Whether the item is the subject of a dispute
                type: boolean
...
    Asset:
      type: object
      additionalProperties: false
      required:
      - value
      - description
      value:
        type: number
        format: decimal
        description: Value of asset
      description:
        type: number
        format: decimal
        description: Description of asset
    Employments:
      type: array
      description: One or more employment income details
      items:
        type: object
        additionalProperties: false
        required:
        - name
        - client_id
        - payments
        description: Employment income detail
        properties:
          name:
            type: string
            description: Identifying name for this employment - e.g. employer's name
          client_id:
            type: string
            description: Client supplied id to identify the employment
          receiving_only_statutory_sick_or_maternity_pay:
            type: boolean
            description: Client is in receipt only of Statutory Sick Pay (SSP) or
              Statutory Maternity Pay (SMP)
          payments:
            "$ref": "#/components/schemas/EmploymentPaymentList"
...
    Applicant:
      type: object
      description: Object describing pertinent applicant details
      required:
      - date_of_birth
      - has_partner_opponent
      - receives_qualifying_benefit
      additionalProperties: false
      properties:
        date_of_birth:
          type: string
          format: date
          example: '1992-07-22'
          description: Applicant date of birth
        employed:
          oneOf:
          - type: boolean
          - type: 'null'
          example: true
        has_partner_opponent:
          type: boolean
          example: false
          description: Applicant has partner opponent (unused in calculation)
        receives_qualifying_benefit:
          type: boolean
          example: false
          description: Applicant receives qualifying benefit
        receives_asylum_support:
          type: boolean
          example: false
          description: Applicant receives section 4 or section 95 Asylum Support
        involvement_type:
          type: string
          enum:
          - applicant
    Assessment:
      type: object
      additionalProperties: false
      required:
      - submission_date
      properties:
        client_reference_id:
          type: string
          example: LA-FOO-BAR
          description: Client's reference number for this application (free text)
        submission_date:
          type: string
          description: Date of the original submission (iso8601 format)
          example: '2022-05-19'
        level_of_help:
          type: string
          enum:
          - certificated
          - controlled
          example: certificated
          description: The level of help required by the client. Defaults to 'certificated'
...
    Dependant:
      type: object
      required:
      - date_of_birth
      - in_full_time_education
      - relationship
      properties:
        date_of_birth:
          type: string
          format: date
          example: '1992-07-22'
        in_full_time_education:
          type: boolean
          example: false
          description: Dependant is in full time education or not
        relationship:
          type: string
          enum:
          - child_relative
          - adult_relative
          example: child_relative
          description: Dependant's relationship to the applicant
        monthly_income:
          "$ref": "#/components/schemas/currency"
          description: Dependant's monthly income
        assets_value:
          "$ref": "#/components/schemas/currency"
          description: Dependant's total assets value
    IrregularIncomePayments:
      type: array
      minItems: 0
      maxItems: 2
      description: One or more irregular payment details
      items:
        type: object
        required:
        - income_type
        - frequency
        - amount
        description: Irregular payment detail
        properties:
          income_type:
            type: string
            enum:
            - student_loan
            - unspecified_source
            description: Identifying name for this irregular income payment
            example: student_loan
          frequency:
            type: string
            enum:
            - monthly
            - annual
            - quarterly
            description: Frequency of the payment received
            example: monthly
          amount:
            "$ref": "#/components/schemas/currency"
...
    ProceedingTypes:
      type: array
      minItems: 1
      description: One or more proceeding_type details
      items:
        type: object
        required:
        - ccms_code
        - client_involvement_type
        properties:
          ccms_code:
            type: string
            enum:
            - DA001
            - DA002
            - DA003
            - DA004
            - DA005
            - DA006
            - DA007
            - DA020
            - SE003
            - SE004
            - SE013
            - SE014
            - SE007
            - SE008
            - SE015
            - SE016
            - SE095
            - SE097
            - SE003A
            - SE004A
            - SE007A
            - SE008A
            - SE013A
            - SE014A
            - SE015A
            - SE016A
            - SE095A
            - SE097A
            - SE101A
            - IM030
            - IA031
            - SE003E
            - SE004E
            - SE007E
            - SE008E
            - SE013E
            - SE014E
            - SE015E
            - SE016E
            - SE096E
            - SE099E
            - SE100E
            - SE101E
            example: DA001
            description: The code expected by CCMS
          client_involvement_type:
            type: string
            enum:
            - A
            - D
            - W
            - Z
            - I
            example: A
            description: The client_involvement_type expected by CCMS
    Property:
      type: object
      required:
      - value
      - outstanding_mortgage
      - percentage_owned
      - shared_with_housing_assoc
      properties:
        value:
          "$ref": "#/components/schemas/currency"
          description: Financial value of the property
        outstanding_mortgage:
          "$ref": "#/components/schemas/currency"
          description: Amount outstanding on all mortgages against this property
        percentage_owned:
          type: number
          format: decimal
          description: Percentage share of the property which is owned by the applicant
          example: 99.99
          minimum: 0
          maximum: 100
        shared_with_housing_assoc:
          type: boolean
          description: Property is shared with a housing association
        subject_matter_of_dispute:
          type: boolean
          description: Property is the subject of a dispute
...
paths:
...
  "/assessments":
    post:
      summary: create assessment
      tags:
      - Assessment
      parameters: []
      responses:
        '200':
          description: successful
        '422':
          description: Unprocessable Entity
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Assessment"
        required: true
...
  "/v6/assessments":
    post:
      summary: create
      tags:
      - Perform assessment with single call
      parameters: []
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                type: object
                required:
                - timestamp
                - result_summary
                - assessment
                - version
                - success
                properties:
                  timestamp:
                    type: string
                  result_summary:
                    type: object
                    required:
                    - overall_result
                    - gross_income
                    - disposable_income
                    - capital
                    properties:
                      overall_result:
                        type: object
                        required:
                        - result
                        - capital_contribution
                        - income_contribution
                        - proceeding_types
                        properties:
                          result:
                            type: string
                            enum:
                            - eligible
                            - ineligible
                            - contribution_required
                          capital_contribution:
                            type: number
                          income_contribution:
                            type: number
                          proceeding_types:
                            type: array
                            items:
                              "$ref": "#/components/schemas/ProceedingTypeResult"
                      gross_income:
                        type: object
                        required:
                        - total_gross_income
                        - combined_total_gross_income
                        - proceeding_types
                        properties:
                          total_gross_income:
                            type: number
                          combined_total_gross_income:
                            type: number
                          proceeding_types:
                            type: array
                            items:
                              "$ref": "#/components/schemas/ProceedingTypeResult"
                      partner_gross_income:
                        type: object
                        required:
                        - total_gross_income
                        properties:
                          total_gross_income:
                            type: number
                      disposable_income:
                        type: object
                        properties:
                          proceeding_types:
                            type: array
                            items:
                              "$ref": "#/components/schemas/ProceedingTypeResult"
                          income_contribution:
                            type: number
                          combined_total_outgoings_and_allowances:
                            type: number
                          total_disposable_income:
                            type: number
                          combined_total_disposable_income:
                            type: number
                          total_outgoings_and_allowances:
                            type: number
                          dependant_allowance:
                            type: number
                          gross_housing_costs:
                            type: number
                          housing_benefit:
                            type: number
                          net_housing_costs:
                            type: number
                          maintenance_allowance:
                            type: number
                          employment_income:
                            type: object
                          partner_allowance:
                            type: number
                      partner_disposable_income:
                        type: object
                        properties:
                          income_contribution:
                            type: number
                          total_disposable_income:
                            type: number
                          total_outgoings_and_allowances:
                            type: number
                          dependant_allowance:
                            type: number
                          gross_housing_costs:
                            type: number
                          housing_benefit:
                            type: number
                          net_housing_costs:
                            type: number
                          maintenance_allowance:
                            type: number
                          employment_income:
                            type: object
                      capital:
                        type: object
                        additionalProperties: false
                        properties:
                          proceeding_types:
                            type: array
                            items:
                              "$ref": "#/components/schemas/ProceedingTypeResult"
                          total_liquid:
                            type: number
                            description: Total value of all client liquid assets in
                              submission
                            format: decimal
                          total_non_liquid:
                            description: Total value of all client non-liquid assets
                              in submission
                            type: number
                            format: decimal
                            minimum: 0.0
                          total_vehicle:
                            description: Total value of all client vehicle assets
                              in submission
                            type: number
                            format: decimal
                          total_property:
                            description: Total value of all client property assets
                              in submission
                            type: number
                            format: decimal
                          total_mortgage_allowance:
                            description: Maxiumum mortgage allowance used in submission.
                              Cases April 2020 will all be set to 999_999_999
                            type: number
                            format: decimal
                          total_capital:
                            description: Total value of all capital assets in submission
                            type: number
                            format: decimal
                          pensioner_capital_disregard:
                            type: number
                            format: decimal
                            description: Cap on pensioner capital disregard for this
                              assessment (based on disposable_income)
                            minimum: 0.0
                          total_capital_with_smod:
                            type: number
                            format: decimal
                            minimum: 0
                            description: Amount of capital with subject matter of
                              dispute deduction applied
                          disputed_non_property_disregard:
                            type: number
                            format: decimal
                            minimum: 0
                            description: Amount of subject matter of dispute deduction
                              applied for assets other than property
                          pensioner_disregard_applied:
                            type: number
                            format: decimal
                            minimum: 0
                            description: Amount of pensioner capital disregard applied
                              to this assessment
                          subject_matter_of_dispute_disregard:
                            type: number
                            format: decimal
                            minimum: 0
                            description: Total amount of subject matter of dispute
                              disregard applied on this submission
                          capital_contribution:
                            type: number
                            format: decimal
                            minimum: 0
                            description: Assessed capital contribution. Will only
                              be non-zero for 'contribution_required' cases
                          assessed_capital:
                            type: number
                            format: decimal
                            minimum: 0
                            description: Amount of assessed client capital. Zero if
                              deductions exceed total capital.
                          combined_assessed_capital:
                            type: number
                            format: decimal
                            minimum: 0
                            description: Amount of assessed capital for both client
                              and partner
                      partner_capital:
                        type: object
                        properties:
                          total_liquid:
                            type: number
                          total_non_liquid:
                            type: number
                          total_vehicle:
                            type: number
                          total_property:
                            type: number
                          total_mortgage_allowance:
                            type: number
                          total_capital:
                            type: number
                          pensioner_capital_disregard:
                            type: number
                          subject_matter_of_dispute_disregard:
                            type: number
                          capital_contribution:
                            type: number
                          assessed_capital:
                            type: number
                  assessment:
                    type: object
                    properties:
                      id:
                        type: string
                      client_reference_id:
                        type: string
                        nullable: true
                        example: ref-11-22
                      submission_date:
                        type: string
                        format: date
                        example: '2022-07-22'
                      level_of_help:
                        type: string
                        enum:
                        - certificated
                        - controlled
                        example: certificated
                        description: The level of representation required by the client
                      applicant:
                        type: object
                      gross_income:
                        type: object
                      disposable_income:
                        type: object
                      capital:
                        type: object
                        additionalProperties: false
                        properties:
                          capital_items:
                            type: object
                            additionalProperties: false
                            required:
                            - liquid
                            - non_liquid
                            - vehicles
                            - properties
                            properties:
                              liquid:
                                type: array
                                items:
                                  "$ref": "#/components/schemas/Asset"
                              non_liquid:
                                type: array
                                items:
                                  "$ref": "#/components/schemas/Asset"
                              vehicles:
                                type: array
                                items:
                                  "$ref": "#/components/schemas/Asset"
                              properties:
                                type: object
                                additionalProperties: false
                                properties:
                                  main_home:
                                    "$ref": "#/components/schemas/PropertyResult"
                                  additional_properties:
                                    type: array
                                    items:
                                      "$ref": "#/components/schemas/PropertyResult"
                  version:
                    type: string
                    enum:
                    - '6'
                  success:
                    type: boolean
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
              - assessment
              - applicant
              - proceeding_types
              properties:
                assessment:
                  "$ref": "#/components/schemas/Assessment"
                applicant:
                  "$ref": "#/components/schemas/Applicant"
                proceeding_types:
                  "$ref": "#/components/schemas/ProceedingTypes"
                capitals:
                  "$ref": "#/components/schemas/Capitals"
                cash_transactions:
                  "$ref": "#/components/schemas/CashTransactions"
                dependants:
                  type: array
                  description: One or more dependants details
                  items:
                    "$ref": "#/components/schemas/Dependant"
                employment_income:
                  "$ref": "#/components/schemas/Employments"
                irregular_incomes:
                  type: object
                  description: A set of irregular income payments
                  required:
                  - payments
                  additionalProperties: false
                  example:
                    payments:
                    - income_type: student_loan
                      frequency: annual
                      amount: 123456.78
                  properties:
                    payments:
                      "$ref": "#/components/schemas/IrregularIncomePayments"
                other_incomes:
                  "$ref": "#/components/schemas/OtherIncomes"
                outgoings:
                  "$ref": "#/components/schemas/OutgoingsList"
                properties:
                  type: object
                  required:
                  - main_home
                  description: A main home and additional properties
                  properties:
                    main_home:
                      "$ref": "#/components/schemas/Property"
                    additional_properties:
                      type: array
                      description: One or more additional properties owned by the
                        applicant
                      items:
                        "$ref": "#/components/schemas/Property"
                regular_transactions:
                  type: array
                  description: Zero or more regular transactions
                  items:
                    "$ref": "#/components/schemas/RegularTransaction"
                state_benefits:
                  type: array
                  description: One or more state benefits receved by the applicant
                    and categorized by name
                  items:
                    "$ref": "#/components/schemas/StateBenefit"
                vehicles:
                  type: array
                  description: One or more vehicles' details
                  items:
                    "$ref": "#/components/schemas/Vehicle"
                partner:
                  type: object
                  required:
                  - partner
                  description: Full information about an applicant's partner
                  example:
                    partner:
                      date_of_birth: '1992-07-22'
                      employed: true
                    irregular_incomes:
                    - income_type: student_loan
                      frequency: annual
                      amount: 101.01
                    employments:
                    - name: A
                      client_id: B
                      payments:
                      - client_id: C
                        date: '1992-07-22'
                        gross: 101.01
                        benefits_in_kind: 0
                        tax: 11
                        national_insurance: 3
                    regular_transactions:
                    - category: benefits
                      operation: credit
                      frequency: three_monthly
                      amount: 101.01
                    state_benefits:
                    - name: D
                      payments:
                      - client_id: E
                        date: '1992-07-22'
                        amount: 101.01
                        flags:
                          multi_benefit: false
                    additional_properties:
                    - value: 500000.01
                      outstanding_mortgage: 999.99
                      percentage_owned: 100
                      shared_with_housing_assoc: false
                      subject_matter_of_dispute: false
                    capitals:
                      bank_accounts:
                      - value: 1.01
                        description: F
                        subject_matter_of_dispute: false
                      non_liquid_capital:
                      - value: 1.01
                        description: G
                        subject_matter_of_dispute: false
                    vehicles:
                    - value: 5000
                      loan_amount_outstanding: 1000
                      date_of_purchase: '2017-01-23'
                      in_regular_use: true
                      subject_matter_of_dispute: false
                  properties:
                    partner:
                      type: object
                      description: The partner of the applicant
                      required:
                      - date_of_birth
                      - employed
                      properties:
                        date_of_birth:
                          type: string
                          format: date
                          example: '1992-07-22'
                          description: Applicant's partner's date of birth
                        employed:
                          type: boolean
                          example: true
                          description: Whether the applicant's partner is employed
                    irregular_incomes:
                      "$ref": "#/components/schemas/IrregularIncomePayments"
                    employments:
                      type: array
                      required:
                      - name
                      - client_id
                      - payments
                      description: One or more employment income details
                      items:
                        type: object
                        description: Employment income detail
                        properties:
                          name:
                            type: string
                            description: Identifying name for this employment - e.g.
                              employer's name
                          client_id:
                            type: string
                            description: Client supplied id to identify the employment
                          payments:
                            "$ref": "#/components/schemas/EmploymentPaymentList"
                    regular_transactions:
                      type: array
                      description: Zero or more regular transactions
                      items:
                        "$ref": "#/components/schemas/RegularTransaction"
                    state_benefits:
                      type: array
                      description: One or more state benefits receved by the applicant's
                        partner and categorized by name
                      items:
                        "$ref": "#/components/schemas/StateBenefit"
                    additional_properties:
                      type: array
                      description: One or more additional properties owned by the
                        applicant's partner
                      items:
                        "$ref": "#/components/schemas/Property"
                    capital_items:
                      "$ref": "#/components/schemas/Capitals"
                    vehicles:
                      type: array
                      description: One or more vehicles' details
                      items:
                        "$ref": "#/components/schemas/Vehicle"
                    dependants:
                      type: array
                      description: One or more dependants details
                      items:
                        "$ref": "#/components/schemas/Dependant"
                explicit_remarks:
                  "$ref": "#/components/schemas/ExplicitRemarks"
        required: true
...